Move IMAP connection into MailRoom::Connection::IMAP
This is in preparation for supporting other mail backends, such as
Microsoft Graph API.
